+++ src/game/state_checkpoint/model_keyword.rs
@@ -47,6 +47,7 @@ pub(super) enum KeywordSnapshot {
     ProtectionFromVampires,
     ProtectionFromWerewolves,
     ProtectionFromCreatures,
+    ProtectionFromMulticolored,
 }
 
 #[derive(Clone, Copy, Debug, Deserialize, Serialize)]